+++ src/components/products/edit-product-form.tsx
@@ -275,7 +275,7 @@ export function EditProductForm() {
                 {/* Buttons */}
                 <div className="mt-2 flex flex-col-reverse gap-3 sm:flex-row sm:justify-end">
                     <Link
-                        href={pathname === '/dashboard/business/entries/create' ? '/dashboard/business/entries' : '/dashboard/business/products'}
+                        href={pathname === '/dashboard/business/inventory/create' ? '/dashboard/business/inventory' : '/dashboard/business/products'}
                         className="bg-transparent border border-white rounded-lg px-3 py-1 text-white flex items-center text-[14px]"
                     >
                         <X className="mr-2 h-4 w-4" />