+++ src/hooks/use-product.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { create, deleteProduct, edit, getAllProducts, getProductById } from "@/lib/api/product";
+import { create, createInBusiness, deleteProduct, edit, getAllProducts, getProductById } from "@/lib/api/product";
 import { CreateProductProps, EditProductProps } from "@/lib/types/product";
 
 
@@ -26,6 +26,18 @@ export function useCreateProductMutation() {
 
     return useMutation({
         mutationFn: (credentials: CreateProductProps) => create(credentials),
+        onSuccess: (_, variables) => {
+            queryClient.invalidateQueries({ queryKey: ["all-product-of-my-businesses", variables.businessId] });
+            queryClient.invalidateQueries({ queryKey: ["all-products"] });
+        },
+    });
+}
+
+export function useCreateProductInBusinessMutation() {
+    const queryClient = useQueryClient();
+
+    return useMutation({
+        mutationFn: (credentials: CreateProductProps) => createInBusiness(credentials),
         onSuccess: (_, variables) => {
             queryClient.invalidateQueries({ queryKey: ["all-product-of-my-businesses", variables.businessId] });
             queryClient.invalidateQueries({ queryKey: ["all-products"] });