+++ src/lib/routes/inventory.ts
@@ -1,7 +1,8 @@
 import { BASIC_ROUTE } from ".";
 
-export const entriesRoutes = {
+export const inventoryRoutes = {
     // getAllEntriesByBusinessId: (businessId: string) => `${BASIC_ROUTE}/entries/business/${businessId}`,
     // createEntry: `${BASIC_ROUTE}/entries`,
     addStockToProduct: (businessId: string, productId: string) => `${BASIC_ROUTE}/inventory/business/${businessId}/product/${productId}/add-stock`,
+    getInventoryByBusinessId: (businessId: string) => `${BASIC_ROUTE}/inventory/business/${businessId}/history`,
 }