+++ src/lib/types/entries.ts
@@ -1,7 +1,0 @@
-export type AddStockToProductProps = {
-    businessId: string;
-    productId: string;
-    quantity: number;
-    unitPrice: number;
-    description: string;
-}